+++ readium/lcp/CHANGELOG.md
@@ -4,11 +4,13 @@ All notable changes to this project will be documented in this file.
 
 **Warning:** Features marked as *alpha* may change or be removed in a future release without notice. Use with caution.
 
-## [Unreleased]
+<!--## [Unreleased]-->
+
+## [2.1.0]
 
 ### Changed
 
-* Upgraded to Kotlin 1.5.21 and Gradle 7.1.1.
+* Upgraded to Kotlin 1.5.31 and Gradle 7.1.1.
 * Migrated to Jetpack Room for the SQLite database storing rights and passphrases (contributed by [@stevenzeck](https://github.com/readium/r2-lcp-kotlin/pull/116)).
     * Note that the internal SQL schema changed. You will need to update your app if you were querying the database manually.
 
@@ -105,4 +107,5 @@ All notable changes to this project will be documented in this file.
 [2.0.0-beta.1]: https://github.com/readium/r2-lcp-kotlin/compare/2.0.0-alpha.2...2.0.0-beta.1
 [2.0.0-beta.2]: https://github.com/readium/r2-lcp-kotlin/compare/2.0.0-beta.1...2.0.0-beta.2
 [2.0.0]: https://github.com/readium/r2-lcp-kotlin/compare/2.0.0-beta.2...2.0.0
+[2.1.0]: https://github.com/readium/r2-lcp-kotlin/compare/2.0.0...2.1.0